requirements.txt
discord.py==2.3.2
gspread==5.11.4
google-auth==2.23.4
matplotlib==3.7.1
pytz==2023.3.post1
numpy==1.24.3
python-dotenv==1.0.0